// src/types/companyPrivacy.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: companyPrivacy
// ====================================================

export interface companyPrivacy_companyPrivacy {
  allPlatform: boolean;
  customerNetwork: boolean;
  supplierNetwork: boolean;
}

export interface companyPrivacy {
  companyPrivacy: companyPrivacy_companyPrivacy | null;
}